Moves TransformComponent into seperate file
